% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-distribution.R
\name{checkDistribution}
\alias{checkDistribution}
\title{Check Distribution}
\usage{
checkDistribution(object)
}
\arguments{
\item{object}{An S4 object.}
}
\value{
TRUE if there are no errors, otherwise the error messages.
}
\description{
Check Distribution
}

