% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineFunctions.R
\name{defineFunctions}
\alias{defineFunctions}
\title{Define functions}
\usage{
defineFunctions(distribution)
}
\arguments{
\item{distribution}{A character value, currently allowed to be one of 
"normal", "student t", "skew normal" or "skew t".}
}
\value{
An object containing the relevant functions.
}
\description{
Define functions
}

