% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-distribution.R
\docType{class}
\name{distribution-class}
\alias{distribution}
\alias{distribution-class}
\title{Distribution Class}
\description{
\describe{
    \item{\code{dev}:}{The function computing the deviance for the
    distribution.}
    \item{\code{gradDev}:}{The function computing the gradient of the
    deviance.}
    \item{\code{paramList2Vec}:}{A function converting a named list of the
    parameters to a vector for optimization.}
    \item{\code{paramVec2List}:}{A function converting a vector of the
    parameters to a named list for easier interpretation/output.}
}
}

