% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNormalDistribution.R
\name{getNormalDistribution}
\alias{getNormalDistribution}
\title{Get Normal Distribution}
\usage{
getNormalDistribution()
}
\value{
A distribution object for the normal distribution.
}
\description{
Get Normal Distribution
}

