% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal.R
\name{multivariateNormal}
\alias{devMN}
\alias{gradDevMN}
\alias{multivariateNormal}
\alias{paramList2VecMN}
\alias{paramVec2ListMN}
\title{Multivariate Normal}
\usage{
devMN(x, params, w = rep(1, nrow(x)))

gradDevMN(x, params)

paramVec2ListMN(paramVec)

paramList2VecMN(paramList)
}
\arguments{
\item{x}{A matrix of observations with one row per observation.}

\item{params}{A vector, typically as created by paramList2Vec called on a 
list object with two elements: xi (a numeric vector) and Omega (a matrix).}

\item{paramVec}{A vector of the parameters of the multivariate distribution.}

\item{paramList}{A list with two elements: xi (a numeric vector) and Omega
(a numeric matrix).}
}
\description{
Functions defining the multivariate normal distribution.  These should be 
used to construct a distribution class as defined in this package.
}
\details{
The deviance function returns the deviance of all the observations with the 
given parameters.  The gradDev function computes the gradient of this 
deviance.  The other two functions are useful for converting parameter
vectors to lists and vice versa, and thus they return the list or vector.
}

