% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_skew_normal.R
\name{multivariateSkewNormal}
\alias{devMSN}
\alias{gradDevMSN}
\alias{multivariateSkewNormal}
\alias{paramList2VecMSN}
\alias{paramVec2ListMSN}
\title{Multivariate Skew Normal}
\usage{
devMSN(x, params)

gradDevMSN(x, params)

paramVec2ListMSN(paramVec)

paramList2VecMSN(paramList)
}
\arguments{
\item{x}{A matrix of observations with one row per observation.}

\item{params}{A vector, typically as created by paramList2Vec called on a 
list object with three elements: xi (a numeric vector) Omega (a matrix),
and alpha (a numeric vector).}

\item{paramVec}{A vector of the parameters of the multivariate distribution.}

\item{paramList}{A list with three elements: xi (a numeric vector), Omega
(a numeric matrix) and alpha (a numeric vector).}
}
\description{
Functions defining the multivariate skew normal distribution.  These should
be used to construct a distribution class as defined in this package.
}
\details{
The deviance function returns the deviance of all the observations with the 
given parameters.  The gradDev function computes the gradient of this 
deviance.  The other two functions are useful for converting parameter 
vectors to lists and vice versa, and thus they return the list or vector.

Note: the multivariate skew normal distribution is implemented as a 
multivariate skew t-distribution with nu parameter equal to 1e8.  This leads
to minimal differences between the two distributions.
}

