% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_t.R
\name{multivariateT}
\alias{devMT}
\alias{gradDevMT}
\alias{multivariateT}
\alias{paramList2VecMT}
\alias{paramVec2ListMT}
\title{Multivariate t}
\usage{
devMT(x, params, w = rep(1, NROW(x)))

gradDevMT(x, params, w = rep(1, NROW(x)))

paramVec2ListMT(paramVec)

paramList2VecMT(paramList)
}
\arguments{
\item{x}{A numeric matrix of observations (one row per observation).}

\item{params}{A vector, typically as created by paramList2Vec called on a 
list object with three elements: xi, Omega, nu.}

\item{paramVec}{A vector of the parameters of the multivariate distribution.}

\item{paramList}{A list with three elements: xi (a vector prodiving the 
center), Omega (a matrix describing the dispersion), and nu (a numeric
providing the heaviness of the tails).}
}
\description{
Functions defining the multivariate t-distribution.
}
\details{
The deviance function returns the deviance of all the observations with the 
given parameters.  The gradDev function computes the gradient of this 
deviance.  The other two functions are useful for converting parameter 
vectors to lists and vice versa, and thus they return the list or vector.
}

