% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal.R
\name{univariateNormal}
\alias{devUN}
\alias{univariateNormal}
\title{Univariate Normal}
\usage{
devUN(x, params)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{params}{A vector, typically as created by paramList2Vec called on a 
list object with two elements: mu and sigma (both numerics).}

\item{paramVec}{A vector of the parameters of the multivariate distribution.}

\item{paramList}{A list with two elements: mean (a numeric) and sigma
(a numeric).}
}
\description{
Functions defining the univariate normal distribution.  These should be 
used to construct a distribution class as defined in this package.
}
\details{
The deviance function returns the deviance of all the observations with the 
given parameters.  The gradDev function computes the gradient of this 
deviance.  The other two functions are useful for converting parameter
vectors to lists and vice versa, and thus they return the list or vector.
}

