% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_skew_t.R
\name{multivariateSkewT}
\alias{devMST}
\alias{gradDevMST}
\alias{multivariateSkewT}
\alias{paramList2VecMST}
\alias{paramVec2ListMST}
\title{Multivariate Skew-t}
\usage{
devMST(x, params, w = rep(1, nrow(x)), fixed.nu = NULL, symmetr = FALSE)

gradDevMST(x, params, w = rep(1, nrow(x)), symmetr = FALSE,
  fixed.nu = NULL)

paramVec2ListMST(paramVec)

paramList2VecMST(paramList)
}
\arguments{
\item{x}{A numeric matrix of observations (one row per observation).}

\item{params}{A vector, typically as created by paramList2Vec called on a 
list object with four elements: xi, Omega, alpha, nu.}

\item{paramVec}{A vector of the parameters of the multivariate distribution.}

\item{paramList}{A list with four elements: xi (a vector prodiving the 
center), Omega (a matrix describing the dispersion), alpha (a vector 
describing the skewness), and nu (a numeric providing the heaviness of the
tails).}
}
\description{
Functions defining the multivariate skew t-distribution.
}
\details{
The deviance function returns the deviance of all the observations with the 
given parameters.  The gradDev function computes the gradient of this 
deviance.  The other two functions are useful for converting parameter 
vectors to lists and vice versa, and thus they return the list or vector.
}

